requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
python-docx==1.1.0
reportlab==4.0.7
bcrypt==4.1.2
pyjwt==2.8.0
web3==6.15.0
aiofiles==23.2.1
slowapi==0.1.9
stripe==7.11.0
coinbase-commerce==2.1.0
python-bitcoinlib==0.12.2
cryptography==41.0.7